i.w.t use spring boot data with jpa/hibernate to access the database

+++ neuralserv-api/src/main/java/org/reggiemcdonald/exception/NumberImageNotFoundException.java
@@ -1,7 +1,7 @@
 package org.reggiemcdonald.exception;
 
 public class NumberImageNotFoundException extends NotFoundException {
-    public NumberImageNotFoundException(int id) {
+    public NumberImageNotFoundException(long id) {
         super(String.format("Number Image ID %d not found", id));
     }
 }
