+++ neuralserv-api/src/main/java/org/reggiemcdonald/exception/TrainingSessionNotFoundException.java
@@ -1,7 +1,7 @@
 package org.reggiemcdonald.exception;
 
 public class TrainingSessionNotFoundException extends NotFoundException {
-    public TrainingSessionNotFoundException(int id) {
+    public TrainingSessionNotFoundException(long id) {
         super(String.format("The training session %d was not found", id));
     }
 }
